rewrote using composition
